Renames for consistency with updated terminology (#123)
Much of the existing code refers to 'entitlements' as properties of the token, or attributes which a token gives permission to.

We've agreed to redefine an 'entitlement' as something which is granted upon an API request. The token is used to obtain the entitlement, but the token is not the entitlement.

This updates the code to better reflect that way of thinking.