upgrade to .net core 2.2 (#126)
The managed projects all currently target dotnet core 2.0. This retargets them to 2.2, as well as updating nuget dependencies to the latest versions.

This also handles a couple of behaviour changes in the latest `System.IdentityModel.Tokens.Jwt` assembly's `ValidateToken` method: [`ValidateIssuerSigningKey` should not `false` if `RequireSignedTokens` is `false`](https://github.com/AzureAD/azure-activedirectory-identitymodel-extensions-for-dotnet/issues/972), and some of the exception messages have changed.