Enhance sestest list-certificates to display certificates appropriate for server authentication (#130)
`sestest` already supports `forsigning` and `forencryption` arguments to `list-certificates`, as well as grouping the output, to easily identify certificates which can be used for signing and encryption.

When running `sestest server` we also require another kind of certificate, i.e. that which can be used for server authentication (to enable clients to communicate via TLS). Identifying certificates which are appropriate for this involves selecting ones which:
- **If** the certificate has an enhanced key usage extension, the usages must include server authentication (OID 1.3.6.1.5.5.7.3.1). Without this the test server will throw an `InvalidOperationException` ("Certificate XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX cannot be used as an SSL server certificate. It has an Extended Key Usage extension but the usages do not include Server Authentication (OID 1.3.6.1.5.5.7.3.1)").
- (optionally) are trusted (and whose certificate chain can be validated).
- have a DNS name which matches the hostname by which requests will be made to the server (probably `localhost`)

This addresses the first two of these by allowing a `forserverauth` argument to `list-certificates`, and displaying two extra sections in the output: one for certificates suitable for server authentication that _can_ be verified, and one for those that can't.